requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
psycopg2-binary==2.9.5
loguru==0.6.0
pydantic==1.10.2